Divided the search into modules
